why did we settle on "latest posts" anyway?
i mean it's not JUST the latest posts. it was always all the posts i've ever made. as the blog grows i might have to paginate it or some shit but for almost a year it's just been the One Post, so i don't think that'll come up anytime soon. and even if so that would make it Page One Of Blog Posts which i guess is technically Page Latest of Blog Posts but like UX wise going from Page Latest to Page 2 would be silly.